processing/requirements.txt
pandas
scikit-learn
snowflake-connector-python
python-dotenv
